requirements.txt
Django==5.1.5
djangorestframework==3.15.2
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.2
sqlparse==0.5.3
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0